pyproject.toml
[tool.poetry]
name = "flask-meteo-in-a-city"
version = "0.1.0"
description = "Get weather forecast for your city"
authors = ["Aleksandr Rol <rolaleksandr75@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0.3"
flask-sqlalchemy = "^3.1.1"
psycopg2-binary = "^2.9.9"
flask-migrate = "^4.0.7"
flask-wtf = "^1.2.1"
python-dotenv = "^1.0.1"
email-validator = "^2.2.0"
flask-login = "^0.6.3"
opencage = "^2.4.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
flake8 = "^7.1.0"
black = "^24.4.2"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
